Hub env post build. Refactoring: ProxyContainerHelper.
